+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 * nostr: add some shorthand constructors for `TagKind::SingleLetter` ([Yuki Kishimoto])
 * nostr: add `Tags` struct ([Yuki Kishimoto])
 * nostr: add `d` tag extraction test from `Tags` ([Yuki Kishimoto])
+* nostr: add `TagStandard::Web` and `TagKind::Web` variants ([Yuki Kishimoto])
 * database: add `Backend::is_persistent` method ([Yuki Kishimoto])
 * database: add `Events` struct ([Yuki Kishimoto])
 * relay-builder: add `LocalRelay` and `RelayBuilder` ([Yuki Kishimoto])